refactor the about us folder system
