+++ components/AboutUs/Hero/index.ts
