+++ components/AboutUs/MainContent/index.ts
