+++ components/AboutUs/ValuesCarousel/ValuesCarousel.module.scss
