+++ components/AboutUs/ValuesCarousel/ValuesCarousel.tsx
