+++ components/AboutUs/ValuesCarousel/index.ts
