+++ pages/about-us.tsx
@@ -2,10 +2,10 @@ import type { NextPage } from 'next';
 import type { AboutUsPage } from 'types/pages/AboutUs';
 import Layout from 'components/Layout';
 import Testimonial from 'components/Testimonial';
-import Hero from 'components/LifeAtTorchbox/Hero';
-import MainContent from 'components/LifeAtTorchbox/MainContent';
-import ValuesCarousel from 'components/LifeAtTorchbox/ValuesCarousel';
-import AtWorkAtPlay from 'components/LifeAtTorchbox/AtWorkAtPlay';
+import Hero from 'components/AboutUs/Hero';
+import MainContent from 'components/AboutUs/MainContent';
+import ValuesCarousel from 'components/AboutUs/ValuesCarousel';
+import AtWorkAtPlay from 'components/AboutUs/AtWorkAtPlay';
 import Metadata from 'components/Metadata';
 import RichText from 'components/RichText';
 import { getAboutUsPage } from 'lib/api';
